optimization: used doubles instead of floats

+++ ImageOps/Blenders/AddBlend.cs
@@ -2,7 +2,7 @@
 {
     public class AddBlend : StandardBlend
     {
-        protected override float Blend(float color1, float color2)
+        protected override double Blend(double color1, double color2)
         {
             return color1 + color2;
         }
