+++ ImageOps/Blenders/AlphaMaskBlend.cs
@@ -4,14 +4,14 @@ namespace ImageOps.Blenders
 {
     public class AlphaMaskBlend : BlendingMethod
     {
-        private readonly Func<PixelColor, float> _alphaSelector;
+        private readonly Func<PixelColor, double> _alphaSelector;
 
         public AlphaMaskBlend(ColorChannel maskChannel)
         {
             _alphaSelector = CreateSelector(maskChannel);
         }
 
-        private Func<PixelColor, float> CreateSelector(ColorChannel maskChannel)
+        private Func<PixelColor, double> CreateSelector(ColorChannel maskChannel)
         {
             switch (maskChannel)
             {
@@ -28,11 +28,11 @@ namespace ImageOps.Blenders
 
         public override PixelColor Blend(PixelColor source, PixelColor mask)
         {
-            float alpha = CalculateAlpha(source.GetAlpha(), _alphaSelector(mask));
+            double alpha = CalculateAlpha(source.GetAlpha(), _alphaSelector(mask));
             return PixelColor.FromFargb(alpha, source.GetRed(), source.GetGreen(), source.GetBlue());
         }
 
-        private float CalculateAlpha(float sourceAlpha, float maskAlpha)
+        private double CalculateAlpha(double sourceAlpha, double maskAlpha)
         {
             return sourceAlpha*maskAlpha;
         }
